ia-redacciones/requirements.txt
fastapi
uvicorn
pyspellchecker
python-dotenv
